# TrafficPatrolProject-bugfix/002implementations/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Trafficpatrol API",
        "description": "Trafficpatrol API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/contents/{id}": {
            "get": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "getContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "updateContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "deleteContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "content-resource"
                ],
                "operationId": "partialUpdateContent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        },
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContentDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminUserDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "updateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUserDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUserDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "public-user-resource"
                ],
                "operationId": "getAllPublicUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{login}": {
            "get": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-resource"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AppUserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "phoneNumber": {
                        "maximum": 1000000000,
                        "minimum": 1000000000,
                        "type": "integer",
                        "format": "int64"
                    },
                    "address": {
                        "type": "string"
                    },
                    "pinCode": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDTO"
                    }
                }
            },
            "ContentDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "contentType": {
                        "type": "string"
                    },
                    "uploadDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "status": {
                        "type": "string"
                    },
                    "appUser": {
                        "$ref": "#/components/schemas/AppUserDTO"
                    }
                }
            },
...
            "UserDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "type": "string"
                    }
                }
            },
...
            "AdminUserDTO": {
                "required": [
                    "login"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "maxLength": 50,
                        "minLength": 1,
                        "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                        "type": "string"
                    },
                    "firstName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "lastName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 254,
                        "minLength": 5,
                        "type": "string"
                    },
                    "imageUrl": {
                        "maxLength": 256,
                        "minLength": 0,
                        "type": "string"
                    },
                    "activated": {
                        "type": "boolean"
                    },
                    "langKey": {
                        "maxLength": 10,
                        "minLength": 2,
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "authorities": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "User": {
                "required": [
                    "activated",
                    "login"
                ],
                "type": "object",
                "properties": {
                    "createdBy": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastModifiedBy": {
                        "type": "string"
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "login": {
                        "maxLength": 50,
                        "minLength": 1,
                        "pattern": "^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$",
                        "type": "string"
                    },
                    "firstName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "lastName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    },
                    "email": {
                        "maxLength": 254,
                        "minLength": 5,
                        "type": "string"
                    },
                    "activated": {
                        "type": "boolean"
                    },
                    "langKey": {
                        "maxLength": 10,
                        "minLength": 2,
                        "type": "string"
                    },
                    "imageUrl": {
                        "maxLength": 256,
                        "minLength": 0,
                        "type": "string"
                    },
                    "resetDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}